+++ controllers/categoriesController.js
@@ -113,11 +113,7 @@ const deleteCategory = async (req, res) => {
             return res.status(200).json({ success: 0, msg: 'No Category in this id!' });
         }
 
-        if (req.files?.image) {
-            img = await imageUpload(req.files.image.name, req.files.image.data);
-            await fs.unlinkSync(found.image);
-            req.body.image = img;
-        }
+        found.image && await fs.unlinkSync(found.image);
 
         const deletedCategory = await Category.deleteOne({ _id: id });
